Extract logics for text_field and chart into respective file
